Add optional workspace-dir input to set where the generation script runs
Running the script from the repo root works fine if your OpenAPI spec is
all in one file, but if it is in separate files it fails because the
main schema references them relative to itself but the script is looking
for them relative to the directory it's running in, which is not
necessarily the same.  For example, if the schemas are in schemas.json
and you include it in the main file with a file reference and their both
in a docs/ directory, you'll get an error like this:

Invalid JSON pointer: ./schemas.json#/components/schemas/something

You can solve this by moving all the OpenAPI files to the root of the
repository where the script is run from but that clutters the repo root.
Alternatively you can run the script from the directory the files are in
so that the relative paths are correct.  For this second option users
need to be able to specify a directory relative to the repo root as the
base path for the OpenAPI schemas if their spec is structured this way.

The new, optional, input `workspace-dir` will be appended to the
`github.workspace` path when setting the `WORKSPACE_DIR` environment
variable that the generation script is already using.  This way the
script will go that directory instead of the repo root and it'll be able
to find the referenced files.