Replace join expression with format
The join expression expects an array as input but it's not clear how you
can make an array yourself during the script, it seems to only support
ones that are made for you by the runtime.  Format supports arbitrary
input though, so it should work.  A trailing or double slash shouldn't
matter if the underlying FS is Linux-like.